% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dataset_bert_pretrained}
\alias{dataset_bert_pretrained}
\title{BERT Pretrained Dataset}
\usage{
dataset_bert_pretrained(x, y = NULL, tokenizer_scheme = NULL, n_tokens = NULL)
}
\arguments{
\item{x}{A data.frame with one or more character predictor columns.}

\item{y}{A factor of outcomes, or a data.frame with a single factor column.
Can be NULL (default).}

\item{tokenizer_scheme}{A character scalar that indicates vocabulary +
tokenizer.}

\item{n_tokens}{An integer scalar indicating the number of tokens in the
output.}
}
\value{
An initialized \code{\link[torch:dataset]{torch::dataset()}}. If it is not yet tokenized, the
\code{tokenize()} method must be called before the dataset will be usable.
}
\description{
Prepare a dataset for pretrained BERT models.
}
