% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{.combine_segments}
\alias{.combine_segments}
\title{Combine a pair of segments}
\usage{
.combine_segments(..., cls_index, sep_index, pad_index, n_tokens)
}
\arguments{
\item{...}{Two lists of tokenized segments. In the future we will support any
number of lists.}

\item{cls_index, sep_index, pad_index}{Named integer vectors for the special
characters.}

\item{n_tokens}{Integer scalar; the number of tokens expected for each
example.}
}
\value{
A list with two components: token_ids and token_type_ids.
}
\description{
Combine a pair of segments
}
\keyword{internal}
