% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{.default_tokenizer}
\alias{.default_tokenizer}
\title{Shortcut to make sure we're using wordpiece}
\usage{
.default_tokenizer(tokenizer)
}
\arguments{
\item{tokenizer}{}
}
\value{
The vocab or wordpiece_tokenize.
}
\description{
Shortcut to make sure we're using wordpiece
}
\keyword{internal}
