% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{.default_vocab}
\alias{.default_vocab}
\title{Shortcut to make sure we're using wordpiece}
\usage{
.default_vocab(vocab)
}
\arguments{
\item{vocab}{}
}
\value{
The vocab or wordpiece_vocab.
}
\description{
Shortcut to make sure we're using wordpiece
}
\keyword{internal}
