% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_pretrained_utils.R
\name{.download_weights}
\alias{.download_weights}
\title{Download and Cache Weights}
\usage{
.download_weights(bert_type = "bert_tiny_uncased", redownload = FALSE)
}
\arguments{
\item{bert_type}{Character; which flavor of BERT to use. See
\code{\link[=available_berts]{available_berts()}} for known models.}

\item{redownload}{Logical; should the weights be downloaded fresh even if
they're cached?}
}
\value{
The parsed weights as a named list.
}
\description{
Download weights for this model to the torchtransformers cache, or load them
if they're already downloaded.
}
\keyword{internal}
