% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{.error_on_tokenizer_mismatch}
\alias{.error_on_tokenizer_mismatch}
\title{Error Helper Function for Mismatches}
\usage{
.error_on_tokenizer_mismatch(new, old)
}
\arguments{
\item{new}{The new value.}

\item{old}{The old value.}
}
\value{
TRUE invisibly.
}
\description{
Error Helper Function for Mismatches
}
\keyword{internal}
