% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{.finalize_attention}
\alias{.finalize_attention}
\title{Finalize Attention Output}
\usage{
.finalize_attention(processed_attention, token_map, pad_token = "[PAD]")
}
\arguments{
\item{processed_attention}{The output from `.process_attention_result`.}

\item{token_map}{The output from `.make_token_map`.}

\item{pad_token}{The token used to pad inputs. Defaults to "[PAD]".}
}
\value{
A tidy version of the attention weights matrix, with one weight value
  per row.
}
\description{
Takes the data frame output from `.process_attention_result`, joins on the
`token_map` information and filters out the padding tokens.
}
\keyword{internal}
