% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{.finalize_bert_tokens}
\alias{.finalize_bert_tokens}
\title{Clean and Return BERT Tokens}
\usage{
.finalize_bert_tokens(tokenized_text, token_types, increment_index)
}
\arguments{
\item{tokenized_text}{A list of integer vectors of token ids.}

\item{token_types}{A list of integer vectors indicating which segment tokens
belong to.}

\item{increment_index}{Logical; if TRUE, add 1L to all token ids to convert
from the Python-inspired 0-indexed standard to the torch 1-indexed
standard.}
}
\value{
An object of class "bert_tokens", which is a list containing a matrix
  of token ids, a matrix of token type ids, and a matrix of token names.
}
\description{
Clean and Return BERT Tokens
}
\keyword{internal}
