% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{.finalize_embeddings}
\alias{.finalize_embeddings}
\title{Finalize Embeddings Output}
\usage{
.finalize_embeddings(processed_embeddings, token_map, pad_token = "[PAD]")
}
\arguments{
\item{processed_embeddings}{The output from `.process_embeddings_result`.}

\item{token_map}{The output from `.make_token_map`.}

\item{pad_token}{The token used to pad inputs. Defaults to "[PAD]".}
}
\value{
A tidy version of the output embeddings, with one embedding vector
  per row.
}
\description{
Takes the data frame output from `.process_embeddings_result`, joins on the
`token_map` information and filters out the padding tokens.
}
\keyword{internal}
