% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_pretrained.R
\name{.get_token_vocab}
\alias{.get_token_vocab}
\title{Look up Token Vocabulary}
\usage{
.get_token_vocab(vocab_name)
}
\arguments{
\item{vocab_name}{Character; the name of the vocabulary.}
}
\value{
The specified token vocabulary.
}
\description{
Given a string representing the name of a token vocabulary, return the
vocabulary.
}
\keyword{internal}
