% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_pretrained.R
\name{.get_tokenizer}
\alias{.get_tokenizer}
\title{Look up Tokenizer Function}
\usage{
.get_tokenizer(tokenizer_name)
}
\arguments{
\item{tokenizer_name}{Character; the name of the tokenization algorithm.}
}
\value{
The function implementing the specified algorithm.
}
\description{
Given a string representing the name of a tokenization algorithm, return the
corresponding tokenization function.
}
\keyword{internal}
