% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_pretrained_utils.R
\name{.get_tokenizer_name}
\alias{.get_tokenizer_name}
\title{Look Up Tokenizer Name}
\usage{
.get_tokenizer_name(tokenizer_scheme)
}
\arguments{
\item{tokenizer_scheme}{A character scalar indicating the tokenizer scheme,
which is a name for a tokenizer plus a vocabulary.}
}
\value{
A character scalar tokenizer name.
}
\description{
Look Up Tokenizer Name
}
\keyword{internal}
