% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_pretrained_utils.R
\name{.get_vocab_name}
\alias{.get_vocab_name}
\title{Look Up Vocabulary Name}
\usage{
.get_vocab_name(tokenizer_scheme)
}
\arguments{
\item{tokenizer_scheme}{A character scalar indicating the tokenizer scheme,
which is a name for a tokenizer plus a vocabulary.}
}
\value{
A character scalar vocab name.
}
\description{
Look Up Vocabulary Name
}
\keyword{internal}
