% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{.infer_segment_index}
\alias{.infer_segment_index}
\title{Figure out Input Segments}
\usage{
.infer_segment_index(token_df, sep_token = "[SEP]", ...)
}
\arguments{
\item{token_df}{A data frame of tokens.}

\item{sep_token}{The token used to demarcate segments. Defaults to "[SEP]".}

\item{...}{Columns to group by. Each group should be a single input sequence.}
}
\value{
The token data frame with a new `segment_index` column.
}
\description{
Given a data frame of input token sequences, add a column giving the segment
index.
}
\keyword{internal}
