% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_weights.R
\name{.load_weights}
\alias{.load_weights}
\title{Load Pretrained Weights into a BERT Model}
\usage{
.load_weights(model, model_name = "bert_base_uncased")
}
\arguments{
\item{model}{a BERT-type model, constructed using `model_bert`.}

\item{model_name}{Character; which flavor of BERT to use. Must be compatible
with `model`!}
}
\value{
The number of model parameters updated. (This is for an error check;
the function is called for side effects.)
}
\description{
There are some hard-to-avoid differences between the variable names in BERT
models constructed using this package and the standard variable names used in
the Huggingface saved weights. This function changes the names from the
Huggingface saves to match package usage.
}
\keyword{internal}
