% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{.make_token_map}
\alias{.make_token_map}
\title{Make Token Map}
\usage{
.make_token_map(tokenized_text, sep_token = "[SEP]")
}
\arguments{
\item{tokenized_text}{A list of tokenized input text. Each element in the
list is expected to be a named integer vector of tokens.}

\item{sep_token}{The token used to demarcate segments. Defaults to "[SEP]".}
}
\value{
A data frame of the input tokens, with explicit index columns.
}
\description{
Given a list of input token sequences, construct a data frame with explicit
index columns.
}
\keyword{internal}
