% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.maybe_alert}
\alias{.maybe_alert}
\title{Filter Alerts through a Verbose Flag}
\usage{
.maybe_alert(verbose = TRUE, ...)
}
\arguments{
\item{verbose}{A logical scalar indicating whether to show the alert.}

\item{...}{Additional parameters passed on to \code{\link[cli:cli_alert]{cli::cli_alert()}}.}
}
\description{
Filter Alerts through a Verbose Flag
}
\keyword{internal}
