% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{.process_attention_result}
\alias{.process_attention_result}
\title{Process Attention}
\usage{
.process_attention_result(bert_model_output)
}
\arguments{
\item{bert_model_output}{The raw output from a BERT model.}
}
\value{
The attention weights from a BERT model, as a data frame.
}
\description{
Takes the raw output from a BERT model and turns the attention weights matrix
into a data frame.
}
\keyword{internal}
