% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_pretrained.R
\name{.process_downloaded_weights}
\alias{.process_downloaded_weights}
\title{Process Downloaded Weights}
\usage{
.process_downloaded_weights(temp_file)
}
\arguments{
\item{temp_file}{The path to the raw downloaded weights.}
}
\value{
The processed weights.
}
\description{
Process Downloaded Weights
}
\keyword{internal}
