% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{.process_embeddings_result}
\alias{.process_embeddings_result}
\title{Process Embeddings}
\usage{
.process_embeddings_result(bert_model_output)
}
\arguments{
\item{bert_model_output}{The raw output from a BERT model.}
}
\value{
The embedding vectors from a BERT model, as a data frame.
}
\description{
Takes the raw output from a BERT model and turns the embeddings into a data
frame.
}
\keyword{internal}
