% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_weights.R
\name{.rename_state_dict_variables}
\alias{.rename_state_dict_variables}
\title{Clean up Parameter Names}
\usage{
.rename_state_dict_variables(state_dict)
}
\arguments{
\item{state_dict}{A state_dict of pretrained weights, probably loaded from a
file.}
}
\value{
The state_dict with the names normalized.
}
\description{
There are some hard-to-avoid differences between the variable names in BERT
models constructed using this package and the standard variable names used in
the Huggingface saved weights. This function changes the names from the
Huggingface saves to match package usage.
}
\keyword{internal}
