% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utilities.R
\name{.standardize_bert_dataset_predictors}
\alias{.standardize_bert_dataset_predictors}
\title{Standardize BERT Dataset Predictors}
\usage{
.standardize_bert_dataset_predictors(x)
}
\arguments{
\item{x}{The input to standardize.}
}
\value{
A data.frame with one or more character columns.
}
\description{
The predictors are coerced to a data.frame of character columns, if possible.
}
\keyword{internal}
