% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{.tokenize_dataset_bert_pretrained}
\alias{.tokenize_dataset_bert_pretrained}
\title{Tokenize Pretrained Bert Datasets}
\usage{
.tokenize_dataset_bert_pretrained(tokenizer_scheme, n_tokens)
}
\arguments{
\item{tokenizer_scheme}{A character scalar that indicates vocabulary +
tokenizer.}

\item{n_tokens}{An integer scalar indicating the number of tokens in the
output.}
}
\value{
TRUE invisibly.
}
\description{
This is really a method of dataset_bert_pretrained.
}
\keyword{internal}
