% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utilities.R
\name{.validate_n_tokens}
\alias{.validate_n_tokens}
\title{Make Sure the Number of Tokens Makes Sense}
\usage{
.validate_n_tokens(n_tokens, allow_null = TRUE)
}
\arguments{
\item{n_tokens}{An integer scalar indicating the number of tokens in the
output.}

\item{allow_null}{A logical scalar indicating whether n_tokens can be NULL.}
}
\value{
n_tokens, validated.
}
\description{
Make Sure the Number of Tokens Makes Sense
}
\keyword{internal}
