% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utilities.R
\name{.validate_tokenizer_metadata}
\alias{.validate_tokenizer_metadata}
\title{Choose Tokenizer Metadata}
\usage{
.validate_tokenizer_metadata(
  bert_type = NULL,
  tokenizer_scheme = NULL,
  n_tokens = NULL,
  allow_null = TRUE
)
}
\arguments{
\item{bert_type}{A bert_type from \code{\link[=available_berts]{available_berts()}} to use to choose the
other properties. If \code{bert_type} and \code{n_tokens} are set, they overrule this
setting.}

\item{tokenizer_scheme}{A character scalar that indicates vocabulary +
tokenizer.}

\item{n_tokens}{An integer scalar indicating the number of tokens in the
output.}

\item{allow_null}{A logical scalar indicating whether the scheme can be NULL.}
}
\value{
A list with elements \code{tokenizer_scheme} and \code{n_tokens}.
}
\description{
Choose Tokenizer Metadata
}
\keyword{internal}
