% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{.validate_tokenizer_scheme}
\alias{.validate_tokenizer_scheme}
\title{Make Sure Tokenizer Schemes are Recognized}
\usage{
.validate_tokenizer_scheme(tokenizer_scheme, allow_null = TRUE)
}
\arguments{
\item{tokenizer_scheme}{A character scalar that indicates vocabulary +
tokenizer.}

\item{allow_null}{A logical scalar indicating whether the scheme can be NULL.}
}
\value{
The scheme, validated.
}
\description{
Make Sure Tokenizer Schemes are Recognized
}
\keyword{internal}
