% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_weights.R
\name{make_and_load_bert}
\alias{make_and_load_bert}
\title{Pretrained BERT Model}
\usage{
make_and_load_bert(model_name = "bert_tiny_uncased")
}
\arguments{
\item{model_name}{Character; which flavor of BERT to use.}
}
\value{
The model with pretrained weights loaded.
}
\description{
Construct a BERT model and load pretrained weights.
}
