% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_pretrained.R
\name{model_bert_pretrained}
\alias{model_bert_pretrained}
\title{Construct a Pretrained BERT Model}
\usage{
model_bert_pretrained(bert_type = "bert_tiny_uncased", redownload = FALSE)
}
\arguments{
\item{bert_type}{Character; which flavor of BERT to use. See
\code{\link[=available_berts]{available_berts()}} for known models.}

\item{redownload}{Logical; should the weights be downloaded fresh even if
they're cached?}
}
\value{
The model with pretrained weights loaded.
}
\description{
Construct a BERT model (using \code{\link[=model_bert]{model_bert()}}) and load pretrained weights.
}
\section{Methods}{

\describe{
\item{\code{initialize}}{Initialize this model. This method is called when the
model is first created.}
\item{\code{forward}}{Use this model. This method is called during training, and
also during prediction. \code{x} is a list of \code{\link[torch:torch_tensor]{torch::torch_tensor()}} values for
\code{token_ids} and \code{token_type_ids}.}
\item{\code{.get_tokenizer_metadata}}{Look up the tokenizer metadata for this
model. This method is called automatically when
\code{\link[=luz_callback_bert_tokenize]{luz_callback_bert_tokenize()}} validates that a dataset is tokenized
properly for this model.}
\item{\code{.load_weights}}{Load the pretrained weights for this model. This
method is called automatically during initialization of this model.}
}
}

