% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_add_norm.R
\name{proj_add_norm}
\alias{proj_add_norm}
\title{Project, Add, and Normalize}
\usage{
proj_add_norm(input_size, output_size, hidden_dropout = 0.1)
}
\arguments{
\item{input_size}{Integer; the size of input tensor.}

\item{output_size}{Integer; the size of output tensor (must match residual).}

\item{hidden_dropout}{Numeric; dropout probability applied after projection.}
}
\description{
Takes in two tensors, an "input" and a "residual". Applies a linear projector
to the input (changing the size to match residual), performs dropout, adds
the result to the residual, then applies layer normalization to the sum.
}
\section{Shape}{


Inputs:
\itemize{
\item input: \eqn{(*, input_size)}
\item residual: \eqn{(*, output_size)}
}

Output:
\itemize{
\item \eqn{(*, output_size)}
}
}

\examples{
in_size <- 4L
out_size <- 3L
model <- proj_add_norm(input_size = in_size, output_size = out_size)
input <- torch::torch_randn(in_size)
residual <- torch::torch_randn(out_size)
model(input, residual)
}
