% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{simplify_bert_token_list}
\alias{simplify_bert_token_list}
\title{Simplify Token List to Matrix}
\usage{
simplify_bert_token_list(token_list)
}
\arguments{
\item{token_list}{A list of vectors. Each vector should have the same length.}
}
\value{
A matrix of tokens. Rows are text sequences, and columns are tokens.
}
\description{
BERT-like models expect a matrix of tokens for each example. This function
converts a list of equal-length vectors (such as a padded list of tokens)
into such a matrix.
}
\examples{
simplify_bert_token_list(
  list(
    1:5,
    2:6,
    3:7
  )
)
}
