% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{tidy_embeddings_output}
\alias{tidy_embeddings_output}
\title{Tidy the Embeddings Output}
\usage{
tidy_embeddings_output(bert_model_output, tokenized_input, pad_token = "[PAD]")
}
\arguments{
\item{bert_model_output}{The output from a BERT model.}

\item{tokenized_input}{A list of tokenized input text. Each element in the
list is expected to be a named integer vector of tokens.}

\item{pad_token}{The token used to pad inputs. Defaults to "[PAD]".}
}
\value{
A tidy version of the output embeddings, with one embedding vector
  per row.
}
\description{
Given the output from a transformer model, construct a tidy data frame.
}
