% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torchtransformers-package.R
\docType{package}
\name{torchtransformers-package}
\alias{torchtransformers}
\alias{torchtransformers-package}
\title{torchtransformers: Transformer Models in Torch}
\description{
Work with transformer models in R using torch.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/macmillancontentscience/torchtransformers}
  \item Report bugs at \url{https://github.com/macmillancontentscience/torchtransformers/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Bratt \email{jonathan.bratt@macmillan.com} (\href{https://orcid.org/0000-0003-2859-0076}{ORCID})

Authors:
\itemize{
  \item Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})
}

Other contributors:
\itemize{
  \item Bedford Freeman & Worth Pub Grp LLC DBA Macmillan Learning [copyright holder]
}

}
\keyword{internal}
