% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_weights.R
\name{available_berts}
\alias{available_berts}
\title{Available BERT Models}
\usage{
available_berts()
}
\value{
A character vector of model names.
}
\description{
List the BERT models that are defined for this package.
}
\examples{
available_berts()
}
