% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_pretrained.R
\name{config_bert}
\alias{config_bert}
\title{BERT Model Parameters}
\usage{
config_bert(
  model_name,
  parameter = c("embedding_size", "n_layer", "n_head", "max_tokens", "vocab_size")
)
}
\arguments{
\item{model_name}{Character scalar; the name of a known BERT model.}

\item{parameter}{Character scalar; the desired parameter.}
}
\value{
Integer scalar; the value of that parameter for that model.
}
\description{
Several parameters define a BERT model. This function can be used to easily
load them.
}
\examples{
config_bert("bert_medium_uncased", "n_head")
}
