% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_weights.R
\name{.load_weights}
\alias{.load_weights}
\title{Load Pretrained Weights into a BERT Model}
\usage{
.load_weights(model, model_name = "bert_base_uncased", redownload = FALSE)
}
\arguments{
\item{model}{A BERT-type model, constructed using `model_bert`.}

\item{model_name}{Character; which flavor of BERT to use. Must be compatible
with `model`!}

\item{redownload}{}
}
\value{
The number of model parameters updated. (This is to enable error
  checks; the function is called for side effects.)
}
\description{
Loads specified pretrained weights into the given BERT model.
}
\keyword{internal}
