% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{.standardize_bert_dataset_outcome}
\alias{.standardize_bert_dataset_outcome}
\title{Standardize BERT Dataset Outcome}
\usage{
.standardize_bert_dataset_outcome(y)
}
\arguments{
\item{y}{A potential outcome variable. Should be a factor, a data.frame with
a single factor column, or NULL.}
}
\value{
A factor or NULL.
}
\description{
Standardize BERT Dataset Outcome
}
\keyword{internal}
