% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_pretrained.R
\name{model_bert_pretrained}
\alias{model_bert_pretrained}
\title{Construct a Pretrained BERT Model}
\usage{
model_bert_pretrained(model_name = "bert_tiny_uncased")
}
\arguments{
\item{model_name}{Character; which flavor of BERT to use. See
\code{\link[=available_berts]{available_berts()}} for known models.}
}
\value{
The model with pretrained weights loaded.
}
\description{
Construct a BERT model (using \code{\link[=model_bert]{model_bert()}}) and load pretrained weights.
}
