% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{simplify_bert_token_list}
\alias{simplify_bert_token_list}
\title{Simplify Token ID List to Matrix}
\usage{
simplify_bert_token_list(token_ids_list, for_torch = TRUE)
}
\arguments{
\item{token_ids_list}{A list of integer vectors.}

\item{for_torch}{Logical; if TRUE, add 1L to all token ids to convert from
the Python-based 0-indexed standard to the torch standard.}
}
\value{
A matrix of token ids.
}
\description{
BERT-like models expect a matrix of tokens for each example. This function
currently supplies a matrix with \code{n_tokens} rows and
\code{length(token_ids_list)} columns, but we plan to transpose those in an
upcoming change to this API.
}
\examples{
simplify_bert_token_list(
  list(
    1:5,
    2:6,
    3:7
  )
)
}
