% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{simplify_bert_token_list}
\alias{simplify_bert_token_list}
\title{Simplify Token ID List to Matrix}
\usage{
simplify_bert_token_list(list_of_integers)
}
\arguments{
\item{list_of_integers}{A list of integer vectors. Each integer should have
the same length.}
}
\value{
A matrix of token ids.
}
\description{
BERT-like models expect a matrix of tokens for each example. This function
currently supplies a matrix with \code{n_tokens} rows and
\code{length(list_of_integers)} columns, but we plan to transpose those in an
upcoming change to this API.
}
\examples{
simplify_bert_token_list(
  list(
    1:5,
    2:6,
    3:7
  )
)
}
